pyproject.toml
[project]
name = "slack-drive-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "slack>=0.0.2",
    "slack-sdk>=3.35.0",
    "slackclient>=2.9.4",
]
